SDC-8004: A few more files with wrong license headers
This patch fixes additional files that we had around with the incorrect
license. As most of the files are either scripts without extension or
weirdness from last update, this change was pretty much done manually.

We're still missing updating several files in the UI module, but since
that was problematical last time, I'll update them in a separate change
(that will be easier to revert if needed).

Change-Id: Ice1456951a71d8fcabdac0e83d3542f522da54af
Reviewed-on: https://review.streamsets.net/11943
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Junko Urata <junko@streamsets.com>