SDC-7870.  Support destination that can write to a Unix named pipe.
In order to use this Destination, you must first create the named pipe
via the mkfifo command.

In the "Named Pipe" tab, enter the full path of the named pipe.

In the destination choose the Data Format for the output data.
All of the formats permitted are one-line, since multi-line
output is probably not what is expected by the client
reading the named pipe.

Data Collector and the named pipe's reader are decoupled and there are some
coordination issues which must be considered.

If Data Collector starts writing to the named pipe before the client is
available to read, the pipeline will remain in STARTING mode until
the reader is available.

If the reader is available, Data Collector will immediately start to
write to the named pipe; if Data Collector writes faster than the reader can
process the data, the pipeline will block when the named pipe is full.

If the pipeline is stopped normally, such as the stop button, the REST API
or via the Pipeline Finisher Executor, the reader will receive an
IOException for Broken Pipe (or C/C++ equavalent).

If the reader is terminated while Data Collector is still writing to
the named pipe, Data Collector will receive the exception, and will throw
a StageException.

Change-Id: I02477506d1d386eab0c7bf20d6b6121a287cb3fb
Reviewed-on: https://review.streamsets.net/12089
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>