SDC-8090: DevDataGenerator origin almost silently swallows exceptions
The DevDataGenerator was first origin to be converted to multithreaded
one and it doesn't propagate all errors as it should have. This patch
changes that to ensure that any execution error will be propagated "up",
thus the pipeline will properly end in error state rather then in
"Pipeline finished" state.

Change-Id: Ic2dd0dcaa511775bb736418a91254a59f9df5464
Reviewed-on: https://review.streamsets.net/12102
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Bob Plotts <bob@streamsets.com>