SDC-8140: Service: Properly delegate methods in PreviewStageLibraryTask
The PreviewStageLibraryTask is a composite StageLibraryTask - it
"inherits" a parent task and injects a custom stage to override origin
if/when needed. Some of the methods were however not properly delegating
up which caused few NPEs in services.

With this patch, the Data Generator stage that rely on service will work
in preview.

Change-Id: Ia9c9730573896f91cc0534fddf23cfdeef352018
Reviewed-on: https://review.streamsets.net/12189
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>