SDC-8151: Service: Make sure that StageRunner is given proper stageClass
While working on SDC-8149, I've noticed that some tests are not calling
the StageRunner with proper stageClass. This is actually incorrect
behavior that might lead to an unexpected outcome (not working ELs for
one).

This patch adds a hard check - it we will now fail if given stageClass
does not have @StageDef() annotation and fixes all the tests that were
passing wrong class.

Change-Id: I9f5e8b87b9249e7c4874579c2abb5b67736395ab
Reviewed-on: https://review.streamsets.net/12200
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>