SDC-8153: Service: Create DataParserServiceImpl
Actual implementation of DataFormatParserService - similarly as the
generator counterpart, the goal for now is to have thin wrapper on top
of existing code. Once we transition all other stages to the service, we
improve the service further. For now it's important that existing data
format library works as-is so that all stages continues working.

Change-Id: I43e755a3af571c39407304b1e93cce0f06668f5c
Reviewed-on: https://review.streamsets.net/12203
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jeff Evans <jeff@streamsets.com>