SDC-8259: Hide retryOnCacheMiss from other processors using CacheConfig
When adding retryOnCacheMiss back in SDC-8175, I made an assertion that
I need to update all processors that are using the LookUpUtils class. It
turns out that some processors are using the CacheConfig without that util
class and hence I missed some processors where the new config needs to be
hidden.

Change-Id: Ibc4d9c61bf6d6539900f484dda4b79e75f84a0d1
Reviewed-on: https://review.streamsets.net/12463
Reviewed-by: Ji Sun Kim <jisun@streamsets.com>