SDC-8273: Remove MemSQL JDBC destination
After doing extensive performance testing, we came to a conclusion that
we need to build a specialed MemSQL destination that utilizes batch load
(via external table concept, pipes) which is tracked by SDC-8241. Since
we will build it soon, it does not make sense to keep the JDBC related
one.

Change-Id: If331152deb5579409bb87ffd43b9b26713d3255d
Reviewed-on: https://review.streamsets.net/12494
Reviewed-by: Ji Sun Kim <jisun@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>