SDC-8354: CDH stage libraries still bundles one jar of Jackson 2.8.9
Back in SDC-8002 we tried to make sure that CDH stage libraries won't pull
in Jackson 2.8.9, but it seems that we were only partially successful. From
some reason even though that we have clear exclusion and that exclusion is
indeed properly shown with {{mvn dependency:tree}}for some reason the maven
assembly plugin is confused and pulls it anyway.

Change-Id: I73f393c0d357b10a24cec78bab540621a9347cea
Reviewed-on: https://review.streamsets.net/12641
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>